Image_Related/PCA+KNN_on_CIFAR10/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cycler==0.10.0            # via matplotlib
django==2.2.4             # via image
image==1.5.27
kiwisolver==1.1.0         # via matplotlib
matplotlib==3.0.3
numpy==1.16.3
pillow==6.0.0             # via image
pyparsing==2.4.0          # via matplotlib
python-dateutil==2.8.0    # via matplotlib
pytz==2019.1              # via django
six==1.12.0               # via cycler, python-dateutil
sqlparse==0.3.0           # via django
